Añadido un entorno virtual, terminados los endpoints, cambios en las propiedades y añadida validación al eliminar

+++ src/main/java/com/drmejia/core/domain/services/implementation/ComprobantServiceImpl.java
@@ -96,6 +96,9 @@ public class ComprobantServiceImpl implements ComprobantService{
 
     @Override
     public void deleteComprobant(@NonNull Long idComprobant){
+        if(!comprobantRepository.existsById(idComprobant)){
+            throw nonExistingComprobant();
+        }
         comprobantRepository.deleteById(idComprobant);
     }
 }
